dom/tests/mochitest/general/test_bug628069_2.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=628069
-->
<head>
  <title>Test for Bug 628069</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=628069">Mozilla Bug 628069</a>
<p id="display"></p>
<div id="content">
  <iframe id="frame" style="height:100px; width:100px; border:0"></iframe>
  <div id="status" style="display: none"></div>
</div>
<pre id="test">
<script type="application/javascript;version=1.7">

/** Test for Bug 628069 **/

gotHashChange = 0;
document.addEventListener("hashChange", function(e) {
  gotHashChange = 1;
  is(e.oldURL, "oldURL");
  is(e.newURL, "newURL");
  is(e.isTrusted, false, "Hashchange event shouldn't be trusted.");
}, true);

let hc = new HashChangeEvent("hashChange", { bubbles: true,
                                             cancelable: false,
                                             oldURL: "oldURL",
                                             newURL: "newURL" });
document.documentElement.dispatchEvent(hc);
is(gotHashChange, 1, 'Document received hashchange event.');

</script>
</body>
</html>
